Use uint64_t for memory addresses
Also fixes an aliasing issue with the explicit template instantiation
in Object.cpp.

+++ include/talvos/DispatchCommand.h
@@ -18,7 +18,7 @@ class Module;
 class Object;
 
 /// Map from Descriptor/Binding pair to an address in memory.
-typedef std::map<std::pair<size_t, size_t>, size_t> DescriptorSet;
+typedef std::map<std::pair<size_t, size_t>, uint64_t> DescriptorSet;
 
 /// This class is used to invoke compute kernels on a virtual device.
 class DispatchCommand
